Renamed test classes in Microsoft.AspNet.Mvc.Xml.Test to end in Test (singular) rather than Tests (plural), according to the engineering guidelines.
